clean imports, reduce expression module, ...
* Interrupt exceptions moved ot of evaluation.py and into expression.py
* Correct return type of Expression.evaluate()
* remove unused imports including those that are imported improperly
  elsewhere
* import from a module only once
