+++ mathics/core/expression.py
@@ -10,6 +10,7 @@ from itertools import chain
 from bisect import bisect_left
 
 from mathics.core.atoms import from_python, Number, Integer
+from mathics.core.interrupt import ReturnInterrupt
 from mathics.core.number import dps
 from mathics.core.convert import sympy_symbol_prefix, SympyExpression
 from mathics.core.symbols import (
@@ -19,10 +20,8 @@ from mathics.core.symbols import (
     Symbol,
     SymbolList,
     SymbolN,
-    SymbolSequence,
     system_symbols,
     ensure_context,
-    strip_context,
 )
 from mathics.core.systemsymbols import SymbolSequence
 
@@ -84,7 +83,7 @@ class BoxError(Exception):
 # ExpressionCache keeps track of the following attributes for one Expression instance:
 
 # time: (1) the last time (in terms of Definitions.now) this expression was evaluated
-#   or (2) None, if the current expression has not yet been evaluatec (i.e. is new or
+#   or (2) None, if the current expression has not yet been evaluated (i.e. is new or
 #   changed).
 # symbols: (1) a set of symbols occuring in this expression's head, its leaves'
 #   heads, any of its sub expressions' heads or as Symbol leaves somewhere (maybe deep
@@ -743,9 +742,12 @@ class Expression(BaseExpression):
         else:
             return self
 
-    def evaluate(self, evaluation) -> typing.Union["Expression", "Symbol"]:
-        from mathics.core.evaluation import ReturnInterrupt
-
+    # When we allow 3.9 only, the return type chould become type[Expression]
+    # See https://adamj.eu/tech/2021/05/16/python-type-hints-return-class-not-instance/
+    def evaluate(self, evaluation) -> typing.Type["Expression"]:
+        """
+        Evaluate until reach a fixed point.
+        """
         if evaluation.timeout:
             return
